Fixes stock update when admin edits order
An admin updating an order should revert the stocks for the order products first by readding them back, and then deducting from stock with the new order product stock value. What it was doing before, it was always deducting when updating the order, so if the order got updated multiple times, it might be deducting a lot from stock.